# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ds-template API",
    "description": "Serves predictions from the latest registered MLflow model, plus run metrics and artifacts for a dashboard.\n\nFeature names are never hardcoded: call `GET /api/predict/schema` to discover what a prediction request needs.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Service health",
        "description": "Report whether the API is serving and whether a model is loaded.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict/schema": {
      "get": {
        "tags": [
          "predict"
        ],
        "summary": "Discover the features a prediction needs",
        "description": "Describe the model's inputs, read from its logged signature.\n\nA frontend calls this to build its prediction form, so that changing dataset\nchanges the form without any code change.",
        "operationId": "predict_schema_api_predict_schema_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict for a single record",
        "description": "Predict for one set of feature values.\n\nRaises:\n    HTTPException: 422 if the submitted features do not match the model\n        signature, listing exactly which names were missing or unknown.",
        "operationId": "predict_api_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict/reload": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Pick up a newly trained model version",
        "description": "Drop the cached model so the newest registered version is loaded.\n\nA server holds one model in memory for the life of the process, so training\na new version while the API runs would otherwise keep serving the old one\nindefinitely. This makes that recoverable without a restart.\n\nRaises:\n    HTTPException: 503 if no model can be loaded after clearing the cache.",
        "operationId": "reload_model_api_predict_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelReloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List runs",
        "description": "List runs for the configured experiment, newest first.\n\nReturns an empty list rather than an error when nothing has been logged, so\na dashboard can render an empty state instead of handling a failure.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get one run",
        "description": "Return a single run with its metrics, parameters, and tags.\n\nRaises:\n    HTTPException: 404 if the run does not exist.",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete a run",
        "description": "Delete one run from MLflow.\n\nThis is a soft delete on MLflow's side (the run moves to its deleted-run\nview rather than disappearing outright), which is enough to drop it out of\nthe dashboard's listing without touching artifact storage here.\n\nRaises:\n    HTTPException: 404 if the run does not exist.",
        "operationId": "delete_run_api_runs__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/artifacts": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List a run's artifacts",
        "description": "List artifacts logged against a run, optionally within a subdirectory.",
        "operationId": "list_artifacts_api_runs__run_id__artifacts_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Subdirectory to list.",
              "default": "",
              "title": "Path"
            },
            "description": "Subdirectory to list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactEntry"
                  },
                  "title": "Response List Artifacts Api Runs  Run Id  Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/artifacts/file": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Download one artifact file",
        "description": "Serve a single artifact file, such as an EDA plot.\n\n``path`` comes from the client, so it is never joined onto the filesystem\ndirectly. It is checked against the run's own artifact listing first, and\nthe resolved file must still sit inside the download directory. Without both\nchecks this endpoint would be a path-traversal hole.\n\nRaises:\n    HTTPException: 404 if the run or artifact does not exist, 400 if the\n        path escapes the artifact tree.",
        "operationId": "get_artifact_file_api_runs__run_id__artifacts_file_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Artifact path, exactly as listed.",
              "title": "Path"
            },
            "description": "Artifact path, exactly as listed."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No model available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactEntry": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Path relative to the run's artifact root."
          },
          "is_dir": {
            "type": "boolean",
            "title": "Is Dir"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          }
        },
        "type": "object",
        "required": [
          "path",
          "is_dir"
        ],
        "title": "ArtifactEntry",
        "description": "One file or directory logged against a run."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation of the failure."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "A failed request. FastAPI uses ``detail`` for its own errors too."
      },
      "FeatureSpecResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Feature name, as the model expects it."
          },
          "mlflow_type": {
            "type": "string",
            "title": "Mlflow Type",
            "description": "MLflow schema type, e.g. 'double'."
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Coarse type for widget selection: number, integer, boolean, string, or datetime."
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "description": "Whether a prediction request must set it."
          },
          "example": {
            "title": "Example",
            "description": "Representative value from the logged input example, suitable for pre-filling a form. Null if unavailable."
          }
        },
        "type": "object",
        "required": [
          "name",
          "mlflow_type",
          "kind",
          "required"
        ],
        "title": "FeatureSpecResponse",
        "description": "One input the model expects, described so a UI can render a control."
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'ok' when the API itself is serving."
          },
          "mlflow_tracking_uri": {
            "type": "string",
            "title": "Mlflow Tracking Uri",
            "description": "Tracking server the API reads."
          },
          "model_available": {
            "type": "boolean",
            "title": "Model Available",
            "description": "Whether a registered model could be loaded. False on a fresh checkout until the training pipeline has run."
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registered model name, when one is loaded."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registered model version, when one is loaded."
          }
        },
        "type": "object",
        "required": [
          "status",
          "mlflow_tracking_uri",
          "model_available"
        ],
        "title": "HealthResponse",
        "description": "Liveness and readiness of the service and its dependencies."
      },
      "ModelReloadResponse": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "reloaded": {
            "type": "boolean",
            "title": "Reloaded",
            "description": "Always true; a failure returns an error instead.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version"
        ],
        "title": "ModelReloadResponse",
        "description": "Which model version is in memory after a reload."
      },
      "PredictRequest": {
        "properties": {
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features",
            "description": "Feature name to value. Call GET /api/predict/schema to discover the expected names and types."
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "PredictRequest",
        "description": "Feature values for a single prediction.\n\nKeys are validated against the model signature when the request is handled,\nnot statically, because the valid keys depend on the trained model."
      },
      "PredictResponse": {
        "properties": {
          "prediction": {
            "title": "Prediction",
            "description": "The predicted label or value."
          },
          "probabilities": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probabilities",
            "description": "Per-class scores, when the estimator exposes predict_proba. Null for regressors and estimators without it."
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "prediction",
          "model_name",
          "model_version"
        ],
        "title": "PredictResponse",
        "description": "The model's answer for one record."
      },
      "PredictSchemaResponse": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureSpecResponse"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version",
          "features"
        ],
        "title": "PredictSchemaResponse",
        "description": "The prediction contract, discovered from the model signature.\n\nA frontend reads this and builds its input form at runtime, so swapping\ndatasets needs no frontend changes."
      },
      "RunDetail": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time",
            "description": "Unix epoch milliseconds, as MLflow reports it."
          },
          "end_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics",
            "description": "Final logged value per metric key. Keys depend on what the pipeline logged, so consumers should render whatever is present."
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Params"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunDetail",
        "description": "One MLflow run with its parameters and tags."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time",
            "description": "Unix epoch milliseconds, as MLflow reports it."
          },
          "end_time": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics",
            "description": "Final logged value per metric key. Keys depend on what the pipeline logged, so consumers should render whatever is present."
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunSummary",
        "description": "One MLflow run, as listed in a dashboard."
      }
    }
  }
}